Factor out the android utils.
